Move taxon functions out into separate module
